Виправив функцію ще раз

+++ lab08/prj/ModulesВірченко/s_calculation.h
@@ -10,7 +10,7 @@ double s_calculation(double x, double y, double z);
 void wind_calculation(int b, std::string& speed, std::string& description);
 
 // Задача 9.2
-void power_and_odds(int n, int m, long long& n_pow_m, long long& m_pow_n, std::vector<int>& odds);
+void power_and_odds(int n, int m, double& n_pow_m, double& m_pow_n, std::vector<int>& odds);
 
 // Задача 9.3
 int binary_analysis(int N);
